requirements.txt
grpcio
grpcio-tools
rasterio==1.3.6
torch==2.0.0
torchvision==0.15.1
matplotlib==3.7.1
tifffile==2022.10.10
numpy==1.26.4
pillow==10.4.0
dill==0.3.8
geopandas==1.0.0
pytest